Allow database to listen on localhost only.
If you set PBS_DATA_SERVICE_HOST, pbs_server assumes you want to
use an external database. If you don't set it, then the database
either listens on all interfaces, which is bad for security, or it
listens where you tell it to, but pbs_server expects it to be the
reverse map of PBS_SERVER, which is not localhost.

So, modify the startup script to start the database before starting
pbs_server.

Includes other localisms.